+++ pygamer/src/SMSimplexID.cpp
@@ -45,6 +45,16 @@ void init_SMSimplexID(py::module& mod){
     );
     vid.def(py::init<>());
 
+    vid.def_property_readonly("dataro",
+        py::overload_cast<>(&SMVertexID::data),
+        R"delim(
+                Access the data stored on the 1-simplex.
+
+                Returns:
+                    :py:class:`Vertex`: Vertex data
+        )delim"
+    );
+
     vid.def("data",
             py::overload_cast<>(&SMVertexID::data),
             py::return_value_policy::reference_internal,
@@ -64,6 +74,24 @@ void init_SMSimplexID(py::module& mod){
                 bool: True if valid.
         )delim"
     );
+    vid.def("indices",
+        &SMVertexID::indices,
+        R"delim(
+            Get the indices of this simplex.
+
+            Returns:
+                list: Indices
+        )delim"
+    );
+    vid.def("cover",
+        &SMVertexID::cover,
+        R"delim(
+            Get the coboundary relations of this simplex.
+
+            Returns:
+                list: Indices of coboundary relations
+        )delim"
+    );
     vid.def("__repr__",
         [](const SMVertexID vid){
             std::ostringstream out;
@@ -72,6 +100,7 @@ void init_SMSimplexID(py::module& mod){
         }
     );
 
+
     // Bindings for SMEdgeID
     py::class_<SMEdgeID> eid(mod, "EdgeID",
         R"delim(
@@ -79,6 +108,15 @@ void init_SMSimplexID(py::module& mod){
 
             This is a token to represent a 2-simplex object. It serves as a
             reference to the actual object.
+        )delim"
+    );
+    eid.def_property_readonly("dataro",
+        py::overload_cast<>(&SMEdgeID::data),
+        R"delim(
+            Access the data stored on the edge.
+
+            Returns:
+                :py:class:`Edge`: Edge data
         )delim"
     );
     eid.def("data", py::overload_cast<>(&SMEdgeID::data),
@@ -99,6 +137,24 @@ void init_SMSimplexID(py::module& mod){
                 bool: True if valid.
         )delim"
     );
+    eid.def("indices",
+        &SMEdgeID::indices,
+        R"delim(
+            Get the indices of this simplex.
+
+            Returns:
+                list: Indices
+        )delim"
+    );
+    eid.def("cover",
+        &SMEdgeID::cover,
+        R"delim(
+            Get the coboundary relations of this simplex.
+
+            Returns:
+                list: Indices of coboundary relations
+        )delim"
+    );
     eid.def("__repr__",
         [](const SMEdgeID eid){
             std::ostringstream out;
@@ -107,6 +163,7 @@ void init_SMSimplexID(py::module& mod){
         }
     );
 
+
     // Bindings for SMFaceID
     py::class_<SMFaceID> fid(mod, "FaceID",
         R"delim(
@@ -114,6 +171,15 @@ void init_SMSimplexID(py::module& mod){
 
             This is a token to represent a 3-simplex object. It serves as a
             reference to the actual object.
+        )delim"
+    );
+    fid.def_property_readonly("dataro",
+        py::overload_cast<>(&SMFaceID::data),
+        R"delim(
+            Access the data stored on the 3-simplex
+
+            Returns:
+                :py:class:`Face`: Face data
         )delim"
     );
     fid.def("data",
@@ -133,6 +199,15 @@ void init_SMSimplexID(py::module& mod){
 
             Returns:
                 bool: True if valid.
+        )delim"
+    );
+    fid.def("indices",
+        &SMFaceID::indices,
+        R"delim(
+            Get the indices of this simplex.
+
+            Returns:
+                list: Indices
         )delim"
     );
     fid.def("__repr__",