+++ tools/blender_addon/src/meshstats.py
@@ -384,13 +384,13 @@ class MeshQualityReportProperties(bpy.types.PropertyGroup):
         description="Select faces with angles less than this criteria")
 
     minCurve = FloatProperty(
-        name="Minimum curvature", default=-1000,
-        description="Lower bound threshold of curvatures"
+        name="Minimum curvature", default=0,
+        description="Lower bound percentile truncation"
         )
 
     maxCurve = FloatProperty(
-        name="Minimum curvature", default=1000,
-        description="Upper bound threshold of curvatures"
+        name="Maximum curvature", default=1000,
+        description="Upper bound percnetile truncation"
         )
 
     minEnergy = FloatProperty(
@@ -403,13 +403,18 @@ class MeshQualityReportProperties(bpy.types.PropertyGroup):
         description="Maximum bound for plotting/thresholding energy"
         )
 
+    niter = IntProperty(
+        name="Smooth Curvature Iterations", default = 5,
+        description="How many iterations of curvature smoothing?"
+        )
 
     def mean_curvature(self, context, report):
         gmesh = blenderToGamer(report)
         if gmesh:
-            curvatures = np.zeros(gmesh.nVertices)
-            for i, vID in enumerate(gmesh.vertexIDs):
-                curvatures[i] = gmesh.getMeanCurvature(vID)
+            curvatures = gmesh.meanCurvature(True, self.niter)
+            # curvatures = np.zeros(gmesh.nVertices)
+            # for i, vID in enumerate(gmesh.vertexIDs):
+            #     curvatures[i] = gmesh.getMeanCurvature(vID)
 
             colors = getColor(curvatures, 'viridis', minV=self.minCurve,
                 maxV=self.maxCurve, percentTruncate=True)
@@ -431,9 +436,7 @@ class MeshQualityReportProperties(bpy.types.PropertyGroup):
     def gaussian_curvature(self, context, report):
         gmesh = blenderToGamer(report)
         if gmesh:
-            curvatures = np.zeros(gmesh.nVertices)
-            for i, vID in enumerate(gmesh.vertexIDs):
-                curvatures[i] = gmesh.getGaussianCurvature(vID)
+            curvatures = gmesh.gaussianCurvature(True, self.niter)
 
             colors = getColor(curvatures, 'viridis', minV=self.minCurve,
                 maxV=self.maxCurve, percentTruncate=True)